resolve NuGet dependency in mpc
